Fix clippy errors

+++ consensus/state_processing/src/per_block_processing/tests.rs
@@ -387,7 +387,7 @@ fn invalid_attestation_wrong_justified_checkpoint() {
             index: 0,
             reason: AttestationInvalid::WrongJustifiedCheckpoint {
                 state: Checkpoint {
-                    epoch: Epoch::from(2 as u64),
+                    epoch: Epoch::from(2_u64),
                     root: Hash256::zero(),
                 },
                 attestation: Checkpoint {
@@ -878,7 +878,7 @@ fn invalid_proposer_slashing_proposal_epoch_mismatch() {
             index: 0,
             reason: ProposerSlashingInvalid::ProposalSlotMismatch(
                 Slot::from(0_u64),
-                Slot::from(128 as u64)
+                Slot::from(128_u64)
             )
         })
     );
